# api/swagger.json
{
  "schemes": [
    "http"
  ],
  "swagger": "2.0",
  "info": {
    "description": "Documentation for My API.",
    "title": "My API.",
    "version": "1.0.0"
  },
  "host": "localhost:8081",
  "paths": {
...
    "/api/v1/tags": {
      "get": {
        "tags": [
          "tags"
        ],
        "summary": "Список всех тегов.",
        "operationId": "listTags",
        "responses": {
          "200": {
            "$ref": "#/responses/getTagResponse"
          },
          "500": {
            "$ref": "#/responses/errorResponse"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "tags"
        ],
        "summary": "Создать новый тег.",
        "operationId": "createTag",
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CreateTagRequest"
            }
          }
        ],
        "responses": {
          "201": {
            "$ref": "#/responses/getTagResponse"
          },
          "400": {
            "$ref": "#/responses/errorResponse"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/api/v1/tags/{id}": {
      "get": {
        "tags": [
          "tags"
        ],
        "summary": "Получить тег по ID.",
        "operationId": "getTag",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "ID",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/getTagResponse"
          },
          "404": {
            "$ref": "#/responses/errorResponse"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      },
      "put": {
        "tags": [
          "tags"
        ],
        "summary": "Обновить тег по ID.",
        "operationId": "updateTag",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "ID",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/getTagResponse"
          },
          "400": {
            "$ref": "#/responses/errorResponse"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      },
      "delete": {
        "tags": [
          "tags"
        ],
        "summary": "Удалить тег по ID.",
        "operationId": "deleteTag",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "ID",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "500": {
            "$ref": "#/responses/errorResponse"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/api/v1/tasks": {
      "post": {
        "tags": [
          "tasks"
        ],
        "summary": "Создать новую задачу.",
        "operationId": "createTask",
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CreateTaskRequest"
            }
          }
        ],
        "responses": {
          "201": {
            "$ref": "#/responses/createTaskResponse"
          },
          "400": {
            "$ref": "#/responses/errorResponse"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/api/v1/tasks/{id}": {
      "get": {
        "tags": [
          "tasks"
        ],
        "summary": "Получить задачу по ID.",
        "operationId": "getTask",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "ID",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/getTaskResponse"
          },
          "500": {
            "$ref": "#/responses/errorResponse"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      },
      "put": {
        "tags": [
          "tasks"
        ],
        "summary": "Обновить задачу.",
        "operationId": "updateTask",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "ID",
            "name": "id",
            "in": "path",
            "required": true
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UpdateTaskRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/getTaskResponse"
          },
          "500": {
            "$ref": "#/responses/errorResponse"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      },
      "delete": {
        "tags": [
          "tasks"
        ],
        "summary": "Удалить задачу.",
        "operationId": "deleteTask",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "ID",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/deleteTaskResponse"
          },
          "500": {
            "$ref": "#/responses/errorResponse"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
...
  },
  "definitions": {
...
    "CreateTagRequest": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "x-go-name": "Name"
        }
      },
      "x-go-package": "github.com/Ozenkol/rbk-go-final/internal/delivery/http/requests"
    },
    "CreateTaskRequest": {
      "description": "CreateTaskRequest represents the request body for creating a task",
      "type": "object",
      "properties": {
        "description": {
          "description": "Task description",
          "type": "string",
          "x-go-name": "Description",
          "example": "Prepare and send offer to client"
        },
        "end_time": {
          "description": "Task end time",
          "type": "string",
          "x-go-name": "EndTime",
          "example": "2023-10-27T11:00:00Z"
        },
        "start_time": {
          "description": "Task start time",
          "type": "string",
          "x-go-name": "StartTime",
          "example": "2023-10-27T10:00:00Z"
        },
        "title": {
          "description": "Task title",
          "type": "string",
          "x-go-name": "Title",
          "example": "Do offer"
        }
      },
      "x-go-package": "github.com/Ozenkol/rbk-go-final/internal/delivery/http/requests"
    },
...
    "Tag": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "x-go-name": "ID"
        },
        "name": {
          "type": "string",
          "x-go-name": "Name"
        }
      },
      "x-go-package": "github.com/Ozenkol/rbk-go-final/internal/domain/tag"
    },
    "Task": {
      "type": "object",
      "properties": {
        "description": {
          "type": "string",
          "x-go-name": "Description"
        },
        "end_time": {
          "type": "string",
          "x-go-name": "EndTime"
        },
        "id": {
          "type": "string",
          "x-go-name": "ID"
        },
        "is_done": {
          "type": "boolean",
          "x-go-name": "IsDone"
        },
        "start_time": {
          "type": "string",
          "x-go-name": "StartTime"
        },
        "title": {
          "type": "string",
          "x-go-name": "Title"
        }
      },
      "x-go-package": "github.com/Ozenkol/rbk-go-final/internal/domain/task"
    },
...
    "UpdateTaskRequest": {
      "description": "UpdateTaskRequest represents the request body for updating a task",
      "type": "object",
      "properties": {
        "description": {
          "description": "Task description",
          "type": "string",
          "x-go-name": "Description",
          "example": "Prepare and send offer to client"
        },
        "end_time": {
          "description": "Task end time",
          "type": "string",
          "x-go-name": "EndTime",
          "example": "2023-10-27T11:00:00Z"
        },
        "is_done": {
          "description": "Task completion status",
          "type": "boolean",
          "x-go-name": "IsDone",
          "example": true
        },
        "start_time": {
          "description": "Task start time",
          "type": "string",
          "x-go-name": "StartTime",
          "example": "2023-10-27T10:00:00Z"
        },
        "title": {
          "description": "Task title",
          "type": "string",
          "x-go-name": "Title",
          "example": "Do offer"
        }
      },
      "x-go-package": "github.com/Ozenkol/rbk-go-final/internal/delivery/http/requests"
    }
  },
  "responses": {
    "createTaskResponse": {
      "description": "",
      "schema": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "x-go-name": "ID"
          }
        }
      }
    },
...
    "deleteTaskResponse": {
      "description": "",
      "schema": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "x-go-name": "Message"
          }
        }
      }
    },
    "errorResponse": {
      "description": "",
      "schema": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "x-go-name": "Error"
          }
        }
      }
    },
...
    "getTagResponse": {
      "description": "",
      "schema": {
        "$ref": "#/definitions/Tag"
      }
    },
    "getTaskResponse": {
      "description": "",
      "schema": {
        "$ref": "#/definitions/Task"
      }
    },
...
  },
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  }
}